# Cache table entries in software to improve Read performance.